move asyncRunner to file and apply some refactoring
